requirements.txt
beautifulsoup4==4.13.3
fastapi==0.115.12
passlib==1.7.4
pydantic==2.11.3
PyJWT==2.10.1
PyJWT==2.10.1
pymongo==4.12.0
selenium==4.31.0
webdriver_manager==4.0.2
